% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laad_data_waterplanten.R
\name{laad_data_waterplanten}
\alias{laad_data_waterplanten}
\title{Laad alle data uit Waterplanten en Waterkwaliteit}
\usage{
laad_data_waterplanten()
}
\value{
Datatabellen zichtbaar in het Global Environment
}
\description{
De functie maakt de tabellen beschikbaar in het Global Environment. De tabellen zijn ook
beschikbaar zonder deze functie uit te voeren.
}
\examples{

\dontrun{
laad_data_waterplanten()
}
}
