% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_data_waterplanten.R
\docType{data}
\name{preferenties}
\alias{preferenties}
\alias{preferentie_klassen}
\alias{preferentie_waarden}
\title{Preferentiewaarden}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24139 rows and 12 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24139 rows and 12 columns.
}
\usage{
preferentie_klassen

preferentie_waarden
}
\description{
Dit zijn alle tabellen met preferentiewaarden van het project waterplanten en waterkwaliteit.
De tabellen zijn in zijn geheel met uitleg te vinden het \emph{Achtergrondrapport Waterplanten en Waterkwaliteit}.
}
\details{
In de tabel \code{preferentie_klassen} is het relatieve voorkomen per klasse opgenomen.
In de tabel \code{preferentie_waarden} zijn de overige preferentiewaarden opgenomen.
}
\seealso{
\href{https://www.stowa.nl/sites/default/files/assets/PUBLICATIES/Publicaties\%202025/STOWA-2025-27\%20Waterplanten\%20en\%20waterkwaliteit/STOWA-2025-27-waterplanten-en-waterkwaliteit.pdf}{Achtergrondrapport Waterplanten en Waterkwaliteit}
}
\keyword{datasets}
