% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZ_data_waterplanten.R
\docType{data}
\name{data_waterplanten}
\alias{data_waterplanten}
\alias{compartments}
\alias{datasets}
\alias{landscape_data}
\alias{landscape_parameters}
\alias{locations}
\alias{methods}
\alias{physicochemical_data}
\alias{physicochemical_parameters}
\alias{samples}
\alias{units}
\alias{vegetation_data}
\alias{zones}
\title{Tabellen}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 71627 rows and 9 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 9 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2561 rows and 12 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 221423 rows and 17 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4324 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24002 rows and 10 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
compartments

datasets

landscape_data

landscape_parameters

locations

methods

physicochemical_data

physicochemical_parameters

samples

units

vegetation_data

zones
}
\description{
Dit zijn alle tabellen van de dataset waterplanten en waterkwaliteit.
Zie ook \url{https://www.stowa.nl/publicaties/rapporten-en-dataset-behorende-bij-het-boek-waterplanten-en-waterkwaliteit}
}
\details{
De databestanden zijn opgenomen in het package als parquet-bestanden onder extdata.
}
\keyword{datasets}
